compiler: extract runtime_iterator
Runtime things are not connected to simulate.
