+++ include/daScript/simulate/runtime_range.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "daScript/simulate/runtime_iterator.h"
 #include "daScript/simulate/simulate.h"
 
 namespace das
